2D Array
We define an hourglass in A to be a subset of values with indices falling in this pattern in 's graphical representation:

a b c
  d
e f g

There are 16 hourglasses in arr and an hourglass sum is the sum of an hourglass' values. Calculate the hourglass sum for every hourglass in , then print the maximum hourglass sum.